improved tray icon, presets
