+++ src/main/resources/messagebundle_de.properties
@@ -53,6 +53,7 @@ fxml.misctab.runlogin=Beim Login starten
 fxml.misctab.eyecare=Augenschoner
 fxml.misctab.brightness=Helligkeit
 fxml.misctab.captureframerate=Bildaufnahme Frequenz
+fxml.misctab.presets=Voreinstellungen
 
 fxml.controltab.gw=für Glow Worm Luciferin
 fxml.controltab.info.graph=Graphen
@@ -86,6 +87,7 @@ tray.icon.start=Start
 tray.icon.stop=Stopp
 tray.icon.settings=Einstellungen
 tray.icon.info=Info
+tray.icon.default=Standard
 tray.icon.exit=Beenden
 
 enum.aspect.ratio.fullscreen=Vollbild
