+++ src/main/resources/messagebundle_en.properties
@@ -53,6 +53,7 @@ fxml.misctab.runlogin=Run at login
 fxml.misctab.eyecare=Eye care
 fxml.misctab.brightness=Brightness
 fxml.misctab.captureframerate=Capture framerate
+fxml.misctab.presets=Presets
 
 fxml.controltab.gw=for Glow Worm Luciferin
 fxml.controltab.info.graph=Info Graphs
@@ -86,6 +87,7 @@ tray.icon.start=Start
 tray.icon.stop=Stop
 tray.icon.settings=Settings
 tray.icon.info=Info
+tray.icon.default=Default
 tray.icon.exit=Exit
 
 enum.aspect.ratio.fullscreen=FullScreen
