+++ src/main/resources/messagebundle_es.properties
@@ -53,6 +53,7 @@ fxml.misctab.runlogin=Iniciar al conectarse
 fxml.misctab.eyecare=Protección de la vista
 fxml.misctab.brightness=Brillo
 fxml.misctab.captureframerate=Velocidad de captura de pantalla
+fxml.misctab.presets=Preajustes
 
 fxml.controltab.gw=for Glow Worm Luciferin
 fxml.controltab.info.graph=Info Graphs
@@ -86,6 +87,7 @@ tray.icon.start=Iniciar
 tray.icon.stop=Parar
 tray.icon.settings=Ajustes
 tray.icon.info=Información
+tray.icon.default=Defecto
 tray.icon.exit=Salir
 
 enum.aspect.ratio.fullscreen=Pantalla completa
