+++ src/main/resources/messagebundle_fr.properties
@@ -53,6 +53,7 @@ fxml.misctab.runlogin=Démarrage Automatique
 fxml.misctab.eyecare=Protection des Yeux
 fxml.misctab.brightness=Luminosité
 fxml.misctab.captureframerate=Fréquence de Capture
+fxml.misctab.presets=Préconfigurations
 
 fxml.controltab.gw=pour Glow Worm Luciferin
 fxml.controltab.info.graph=Graphs D'infos
@@ -86,6 +87,7 @@ tray.icon.start=Démarrer
 tray.icon.stop=Éteindre
 tray.icon.settings=Réglages
 tray.icon.info=Info
+tray.icon.default=Défaut
 tray.icon.exit=Quitter
 
 enum.aspect.ratio.fullscreen=Plein-Écran
