+++ src/main/resources/messagebundle_hu.properties
@@ -53,6 +53,7 @@ fxml.misctab.runlogin=Futtatás indításkor
 fxml.misctab.eyecare=Eye care
 fxml.misctab.brightness=Fényerő
 fxml.misctab.captureframerate=Képvételezésés sebessége
+fxml.misctab.presets=Előbeállítások
 
 fxml.controltab.gw=Glow Worm Luciferin-nek
 fxml.controltab.info.graph=Infó grafikonok
@@ -86,6 +87,7 @@ tray.icon.start=Start
 tray.icon.stop=Stop
 tray.icon.settings=Beállítások
 tray.icon.info=Infó
+tray.icon.default=Alapértelmezett
 tray.icon.exit=Kilép
 
 enum.aspect.ratio.fullscreen=Teljes képernyő
