+++ src/main/resources/messagebundle_it.properties
@@ -53,6 +53,7 @@ fxml.misctab.runlogin=Avvio automatico
 fxml.misctab.eyecare=Affaticamento occhi
 fxml.misctab.brightness=Luminosità
 fxml.misctab.captureframerate=Framerate
+fxml.misctab.presets=Presets
 
 fxml.controltab.gw=for Glow Worm Luciferin
 fxml.controltab.info.graph=Info Grafici
@@ -86,6 +87,7 @@ tray.icon.start=Start
 tray.icon.stop=Stop
 tray.icon.settings=Impostazioni
 tray.icon.info=Info
+tray.icon.default=Predefinito
 tray.icon.exit=Esci
 
 enum.aspect.ratio.fullscreen=Schemo intero
