+++ src/main/resources/messagebundle_ru.properties
@@ -53,6 +53,7 @@ fxml.misctab.runlogin=Запуск при старте
 fxml.misctab.eyecare=Ночной режим
 fxml.misctab.brightness=Яркость
 fxml.misctab.captureframerate=Частота кадров
+fxml.misctab.presets=Пресеты
 
 fxml.controltab.gw=для Glow Worm Luciferin
 fxml.controltab.info.graph=Информационные графики
@@ -86,6 +87,7 @@ tray.icon.start=Запустить
 tray.icon.stop=Остановить
 tray.icon.settings=Настройки
 tray.icon.info=Информация
+tray.icon.default=По умолчанию
 tray.icon.exit=Выход
 
 enum.aspect.ratio.fullscreen=Полный экран
